Add TestSnapshot test suite
